refactor: 리뷰 API 통합 및 구조 개선
